Fix clientes Controllers
