+++ controladores/controladorClientes.js
@@ -30,22 +30,26 @@ function obtenerClientesPorId(req, res) {
     }
 }
 
-function crearNuevoCliente(req, res) {
+function crearNuevoCliente(req, res){
     let body = req.body;
     let nombre = body.nombre;
-    let telefono = body.telefono;
+    let telefono = body.telefono;    
 
-    conexion.query("INSERT INTO client (nombre, telefono) VALUES (?,?)", [nombre, telefono],
-        function (err, result) {
-            if (err) return console.log("Hubo un error en el insert de cliente", err.message);
-            res.send(result);
-        });
+    if(!nombre || typeof(nombre)!=='string') return res.status(422).json({ 'message': 'contenido invalido', 'cause': 'no hay nombre o el formato es incorrecto' });
+    
+    if(!telefono || typeof(telefono)!=='number') return res.status(422).json({ 'message': 'contenido invalido', 'cause': 'no hay telefono o el formato es incorrecto' });
 
+    conexion.query("INSERT INTO client (nombre, telefono) VALUES (?,?)", [nombre, telefono], 
+    function(err, result){
+        if(err) return console.log("Hubo un error en el insert de cliente", err.message);         
+        res.send(result);
+    });
+    
 }
 
 module.exports = {
-    obtenerClientes: obtenerClientes,
-    obtenerClientesPorId: obtenerClientesPorId,
-    crearNuevoCliente: crearNuevoCliente
+    obtenerClientes,
+    obtenerClientesPorId,
+    crearNuevoCliente
 };
